requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
blas=1.0=mkl
brotlipy=0.7.0=py37h540881e_1004
bzip2=1.0.8=h7f98852_4
ca-certificates=2022.6.15=ha878542_0
certifi=2022.6.15=py37h89c1867_0
cffi=1.14.6=py37hc58025e_0
charset-normalizer=2.1.0=pyhd8ed1ab_0
cryptography=37.0.2=py37h38fbfac_0
cudatoolkit=11.6.0=hecad31d_10
distro=1.7.0=pypi_0
ffmpeg=4.3=hf484d3e_0
freetype=2.10.4=h0708190_1
func-timeout=4.3.5=pypi_0
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
idna=3.3=pyhd8ed1ab_0
intel-openmp=2021.4.0=h06a4308_3561
jpeg=9e=h166bdaf_1
lame=3.100=h7f98852_1001
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.17=h166bdaf_0
libpng=1.6.37=h21135ba_2
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.2.0=hf544144_3
libwebp-base=1.2.2=h7f98852_1
lz4-c=1.9.3=h9c3ff4c_1
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h402132d_0
mkl_fft=1.3.1=py37h3e078e5_1
mkl_random=1.2.2=py37h219a48f_0
ncurses=6.3=h5eee18b_3
nettle=3.6=he412f7d_0
networkx=2.6.3=pypi_0
numpy=1.18.4=pypi_0
numpy-base=1.21.5=py37ha15fc14_3
olefile=0.46=pyh9f0ad1d_1
openh264=2.1.1=h780b84a_0
openjpeg=2.4.0=hb52868f_1
openssl=1.1.1o=h166bdaf_0
pillow=8.2.0=py37h4600e1f_1
pip=21.2.2=py37h06a4308_0
psutil=5.9.1=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pysocks=1.7.1=py37h89c1867_5
python=3.7.13=h12debd9_0
python_abi=3.7=2_cp37m
pytorch=1.12.0=py3.7_cuda11.6_cudnn8.3.2_0
pytorch-mutex=1.0=cuda
pyyaml=6.0=pypi_0
readline=8.1.2=h7f8727e_1
requests=2.28.1=pyhd8ed1ab_0
setuptools=61.2.0=py37h06a4308_0
shapely=1.8.2=pypi_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.38.5=hc218d9a_0
tk=8.6.12=h1ccaba5_0
torchaudio=0.12.0=py37_cu116
torchvision=0.13.0=py37_cu116
typing_extensions=4.3.0=pyha770c72_0
urllib3=1.26.10=pyhd8ed1ab_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7f8727e_1
zlib=1.2.12=h7f8727e_2
zstd=1.5.0=ha95c52a_0
